remove troublesome userAgent module
