+++ package.json
@@ -103,7 +103,7 @@
     "watch": "jung -D '\\.npm|.git|node_modules|build|tools|tests' --run -- npm run build:all",
     "cdn:build": "npm run cdn:build:prod",
     "cdn:build:extension": "PUBLISH=EXTENSION rm -rf ./build && npm run cdn:webpack && npm run cdn:cleanup",
-    "cdn:build:local": "npm run cdn:webpack && npm run cdn:cleanup && npm run build:browser-tests",
+    "cdn:build:local": "npm run cdn:webpack && npm run cdn:cleanup",
     "cdn:build:prod": "npm run cdn:fresh-build && PUBLISH=PROD npm run cdn:webpack && npm run cdn:cleanup && npm run cdn:clone",
     "cdn:build:dev": "npm run cdn:fresh-build && PUBLISH=DEV npm run cdn:webpack && npm run cdn:cleanup",
     "cdn:build:current": "npm run cdn:fresh-build && PUBLISH=CURRENT npm run cdn:webpack && npm run cdn:cleanup",
