+++ src/common/config/state/runtime.js
@@ -1,4 +1,3 @@
-import * as userAgent from '../../util/user-agent'
 import { getModeledObject } from './configurable'
 import { gosNREUMInitializedAgents } from '../../window/nreum'
 import { globalScope } from '../../util/global-scope'
@@ -23,7 +22,6 @@ const model = {
   releaseIds: {},
   session: undefined,
   xhrWrappable: typeof globalScope.XMLHttpRequest?.prototype?.addEventListener === 'function',
-  userAgent,
   version: VERSION
 }
