+++ src/features/jserrors/aggregate/index.js
@@ -20,13 +20,13 @@ import { globalScope } from '../../../common/util/global-scope'
 import { FEATURE_NAME } from '../constants'
 import { drain } from '../../../common/drain/drain'
 import { FEATURE_NAMES } from '../../../loaders/features/features'
-import { FeatureBase } from '../../utils/feature-base'
+import { AggregateBase } from '../../utils/aggregate-base'
 
 /**
  * @typedef {import('./compute-stack-trace.js').StackInfo} StackInfo
  */
 
-export class Aggregate extends FeatureBase {
+export class Aggregate extends AggregateBase {
   static featureName = FEATURE_NAME
   constructor (agentIdentifier, aggregator) {
     super(agentIdentifier, aggregator, FEATURE_NAME)
