+++ src/features/utils/feature-base.js
@@ -1,7 +1,5 @@
-import { getInfo, isConfigured, getRuntime } from '../../common/config/config'
+import { getRuntime } from '../../common/config/config'
 import { ee } from '../../common/event-emitter/contextual-ee'
-import { configure } from '../../loaders/configure/configure'
-import { gosCDN } from '../../common/window/nreum'
 
 export class FeatureBase {
   constructor (agentIdentifier, aggregator, featureName) {
@@ -19,34 +17,5 @@ export class FeatureBase {
      * @type {boolean}
      */
     this.blocked = false
-
-    this.checkConfiguration()
-  }
-
-  /**
-   * Checks for additional `jsAttributes` items to support backward compatibility with implementations of the agent where
-   * loader configurations may appear after the loader code is executed.
-   */
-  checkConfiguration () {
-    // NOTE: This check has to happen at aggregator load time
-    if (!isConfigured(this.agentIdentifier)) {
-      let jsAttributes = { ...gosCDN().info?.jsAttributes }
-      try {
-        jsAttributes = {
-          ...jsAttributes,
-          ...getInfo(this.agentIdentifier)?.jsAttributes
-        }
-      } catch (err) {
-        // do nothing
-      }
-      configure(this.agentIdentifier, {
-        ...gosCDN(),
-        info: {
-          ...gosCDN().info,
-          jsAttributes
-        },
-        runtime: getRuntime(this.agentIdentifier)
-      })
-    }
   }
 }
